Set up Burushaski-to-English translation model and chatbot framework
Burushaski to Chatbot
This repository focuses on developing a chatbot system that facilitates Burushaski-to-English translation. It utilizes Natural Language Processing (NLP) techniques to train a model capable of understanding and generating responses in Burushaski, a language spoken in the Gilgit-Baltistan region, and translating them to English. The project aims to improve communication for speakers of Burushaski and support language preservation.